backend/pyproject.toml
[project]
name = "secucert-backend"
version = "0.1.0"
description = "Security Certificate Monitor Backend"
requires-python = ">=3.9"
dependencies = [
    "flask>=2.2.5,<3.0",
    "flask-cors>=4.0.0",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.0",
    "apscheduler>=3.10.4",
    "pydantic>=2.0.0",
    "faker>=20.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-flask>=1.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
